Bug #23762382: INSERT VALUES QUERY WITH JOIN IN A SELECT                CAUSES INCORRECT BEHAVIOR
Issue:
------
Optimizer is unable to handle table/column permissions when
the INSERT statement has the following features:
1) The VALUES-list of the INSERT statement contains a
   SELECT statement with a join.
2) And it is not in the first row of the VALUES list.

Currently, the steps to prepare an INSERT statement are:
1) The first row in the VALUES-list goes for fix_fields
   from Sql_cmd_insert_base::mysql_prepare_insert.
  a) This calls the setup_fields which initializes
     privileges using the Column_privilege_tracker.
  b) When resolving a column, optimizer checks the
     privileges.
2) Local transforms are applied for the whole INSERT
   statement.
  a) Tries to simplify_joins and ultimately does fix_fields
     if necessary.
3) The rest of the rows in the VALUES-list go for
   fix_fields from Sql_cmd_insert::mysql_insert.
  a) This calls the setup_fields which initializes
     privileges using the Column_privilege_tracker.
  b) When resolving a column, optimizer checks the
     privileges.

The current problem occurs in Step 2a. Since the
setup_fields hasn't been called, the privileges for the
column haven't been set. And during resolution of the
columns there is a problem.

Solution:
---------
Ideally, local transforms should be applied only after
fix_fields have been completed through setup_fields.

With this patch, all the rows in the VALUES-list are
resolved from Sql_cmd_insert_base::mysql_prepare_insert.

The problem has been addressed in mysql-trunk by WL#5094.
But this is not a backport, since a significant part of the
code has been re-factored in the WL.