Bug#24432199 MTR'S INTERNAL CHECK OF THE TEST CASE 'MAIN.MYSQL_UPGRADE' FAILED
When a sys schema entity is re-created, the entry in mysql.user and
mysql.tables_priv tables is done again and this changes the timestamps
in these tables. The test 'main.mysql_upgrade' does the above operation
and hence, the checksum of the two tables differ and MTR's check of the
testcase fails. A backup of the tables mysql.user and mysql.tables_priv
is taken and restored after the above scenario to resolve the issue.

Reviewed-by: Mark Leith <mark.leith@oracle.com>
RB: 14348