Bug #24823885: PERFORMANCE REGRESSION WHEN CHANGING CHARACTER SET TO UTF8MB4
Add a fastpath for simple (non-tailored) UCA 9.0.0 collations,
where as long as the string is pure ASCII, we can skip a lot of
the checks and lookups normally required. Once we hit a non-ASCII
character, we fall back to the regular path, but return if we
see ASCII characters again. (This ensures a single non-ASCII
character, such as an emoji, won't make us fall off the fast path
entirely.)

This is based on an idea and initial implementation from Xing,
but adjusted so that it is somewhat more generic.

Microbenchmark results (64-bit, Skylake 3.4 GHz, GCC 6.1):

  BM_SimpleUTF8MB4          1329 -> 363 ns/iter  [+266.1%]
  BM_MixedUTF8MB4            603 -> 321 ns/iter  [ +87.9%]
  BM_MixedUTF8MB4_AS_CS     2037 -> 982 ns/iter  [+107.4%]

sysbench goes from 5429 -> 7431 tps (+36.9%).

Change-Id: I4cc22e7141ba7b3ddd99da97bf32ef30156fb28b