WL#9457: Deprecate support for non-native partitioning.
This WL inplements the following changes:

1. The bootstrap implementation in the server is extended with
   support for executing a single query.
2. While restarting the server, an I_S query is executed to
   trigger a scan of all .FRM files in the server.
3. The query in 2. is skipped if the command line option
   '--disable-partition-engine-check' is submitted.
4. When opening an .FRM file, a deprecation warning is issued
   if the table in question is using the non-native partition
   engine. Before the server is open to external connections,
   the warning is printed to the error log. Afterwards, the
   warning is returned to the client.
5. The warning in 3. is suppressed for I_S queries, CHECK
   TABLE, etc. Instead, to ensure repeatable results, the
   warning is issued in the processing of the meta data.
6. Add test wrappers for tests being run with different default
   storage engines to keep results separate for engines not
   supporting native partitioning.
7. Test changes due to deprecation warnings. Record new result
   file and slight test changes to make results predictable and
   repeatable.

Item 4 above means that for the I_S query executed during
server restart, there will be a warning in the error log for
each non-natively partitioned table.

For DDL and DML statements accessing a non-natively partitioned
table, a warning will be issued if the statement execution
makes an .FRM file be opened. This is done to keep the amount
of warnings down for DML statements. For CREATE and ALTER table,
a warning will always be issued.