Bug#24949179: Handle simultaneous DD cache misses with object eviction.
In the shared DD cache, when several threads tried to acquire the same
non-present DD object simultaneously, only the first would handle the cache
miss by fetching the object from disk. When the first thread was putting the
object into the shared cache, the other threads (waiting for a condition
variable) were broadcasted, and one of them would proceed with the assumption
that the object was indeed in the shared cache. However, this might not be
the case if the cache is under sufficient stress that the object is evicted
from the cache after the first thread has put it there, but before the second
thread tries to get and return it.

The fix is to let a thread being awakened after waiting for a cache miss
being processed continue and try to fetch the object from disk if it was not
present in the shared cache, i.e., treat a missing object as a cache miss in
the normal way.