Bug #24823885: PERFORMANCE REGRESSION WHEN CHANGING CHARACTER SET TO UTF8MB4
When writing a big-endian 16-bit weight to the buffer, do it in one go
(using htons and memcpy) rather than shifting and masking ourselves;
seemingly GCC doesn't manage to combine the two writes. (The htons()
is rewritten to a 16-bit rotate-by-8, which is efficient although it
is contrary to Intel's recommendation to use XCHG AL, AH.)

Microbenchmark results (Skylake 3.4 GHz, opt, GCC 6.2):

  BM_SimpleUTF8MB4          375 ->  309 ns/iter  [+21.4%]
  BM_MixedUTF8MB4           316 ->  307 ns/iter  [ +1.3%]
  BM_MixedUTF8MB4_AS_CS    1041 ->  982 ns/iter  [ +6.1%]

sysbench goes from 10825 -> 11136 tps (+2.9%).

Change-Id: Ifa54a02456dcaec808bef2fc4fb5c6d2158133d7