Bug#23581389: SEMI-SYNC CAN'T RUN NORMALLY WHEN MANY CONNECTIONS
Problem & Analysis:
Semisync Ack thread uses select() call to check if sockets are
available for read. But select() uses FD_SET which is limited
by FD_SETSIZE(1024 by default).

Select() man page says:
#######################
"An fd_set is a fixed size buffer.  Executing FD_CLR() or FD_SET() with a
value of fd that is negative or is equal to or  larger  than
 FD_SETSIZE will result in undefined behavior."
########################

Fix: Semisync ack thread is changed to use poll() instead of select().
On some systems, poll may not be available, hence the code is wrapped
around HAVE_POLL define.
On the systems that still use select(), Ack receive thread will generate
an error and semi sync will be switched off. Windows systems is exception
case because on windows this limitation does not exists.