Bug#24968735 - MODIFYVIEW `STMT` FIELD SHOULDNT BE MARKED AS REQUIRED
Description
===========
`stmts` field defined in `Mysqlx.Crud.ModifyView` message was marked as
required which limits future modification/features. This field must be
marked as optional.

Fix
===
The definition of "protobuf" field was changed to "optional". When
X Plugin detects that the field is missing it going to generate an
error.

RB: 14436
Reviewed-by Andrzej Religa <andrzej.religa@oracle.com>