Bug#24364448: assert m_opr_handle != null in ha_innobase_inmem::keys_to_use_for_scanning
It turns out calling this function from test_if_cheaper_ordering() is
both unnecessary and does not give the desired behavior. Hence, this
bug can be avoided by removing this call.

handler::keys_to_use_for_scanning() returns the set of indexes that
can be used to provide order.  In other words, it returns the set of
B-tree indexes, while hash indexes is not included.  This should not
be necessary to check since the indexes considered by
test_if_cheaper_ordering() is based on Field::part_of_sortkey.  Hence,
hash indexes have already been excluded at this point.

Also, the code in test_if_cheaper_ordering() does not achieve what is
intended.  The comment says: "If not used with LIMIT, only use keys if
the whole query can be resolved with a key; This is because filesort()
is usually faster than retrieving all rows through an index."  In
other words, one only wants to switch to full index scan if the index
is covering.  However, the set of indexes considered is the union of
covering indexes and the indexes returned by keys_to_use_for_scanning(),
which in InnoDB's case will be all indexes.   (While MyISAM's use the
default implementation and returns no indexes.)

Other comments in test_if_cheaper_ordering() are in conflict with the
comment cited above.  For example: "Don't use an index scan with ORDER
BY without limit.  For GROUP BY without limit always use index scan if
there is a suitable index."  Due to the effect of the code discussed
above, this has been the effect for InnoDB while for MyISAM full index
scan for GROUP BY has only been used when the index has been covering.

The fix is to remove the discussed code.  This will not make any
difference for InnoDB tables since all indexes were already
considered.  For MyISAM tables, full index scan may now be used
instead of full table scan + filesort even when index is not covering.

Details:

  sql/sql_select.cc
    Remove the code that limits the indexes that may be considered by
    test_if_cheaper_ordering()
    Remove local variable keys since parameter usable_keys may be used

  mysql-test/t/derived.test
  mysql-test/r/derived.result
    Reduced test case from bug report
    Note that this test case does not fail on trunk; only in WL#8117 branch

  mysql-test/t/heap_btree.test
  mysql-test/r/heap_btree.result
    Added a test case that verifies that btree index in heap engine may
    be used for ordering

  mysql-test/t/heap_hash.test
  mysql-test/r/heap_hash.result
    Added a test case that verifies that hash index is not used for ordering

  mysql-test/r/group_by.result
    Result of a query with not full group by changes since rows is
    accessed in different order when full index scan is used.

  mysql-test/r/group_min_max.result
    Switch from table scan with sorting to full index scan that avoids
    sorting

  mysql-test/r/key.result
    Switch from table scan with sorting to full index scan that avoids
    sorting

  mysql-test/r/select_found.result
    Switch from table scan with sorting to full index scan that avoids
    sorting
    Distinct query with limit gives different result since rows is
    accessed in different order when full index scan is used.

  mysql-test/r/subquery_mat.result
  mysql-test/r/subquery_mat_all.result
  mysql-test/r/subquery_mat_none.result
    Switch from table scan with sorting to full index scan that avoids
    sorting

  mysql-test/suite/gcol/r/gcol_keys_myisam.result
  mysql-test/suite/gcol/r/gcol_select_myisam.result
    Switch from table scan with sorting to full index scan that avoids
    sorting