Bug#24384561 - 5.7.14 COMPLAINS ABOUT WRONG                SLAVE_MASTER_INFO AFTER UPGRADE FROM 5.7.13
DESCRIPTION
===========
After upgrading from 5.6.21 to 5.7.13 and then to 5.7.14,
the mysql.slave_master_info table reports an error that
the fields 'Channel_name' and 'Tls_version' are not in the
expected order.

ANALYSIS
========
It is expected that the column Tls_version comes after
column Channel_name in the slave_master_info table.
When upgrade happens from 5.6.21  directly to the 5.7.13+
release, these fields which were not present earlier are
added in a correct order.

The columns Channel_name and Tls_version are present in an
incorrect order in 5.7.13. So, when the upgrade happens
from 5.7.13 to 5.7.13+, the upgrade tool does not
correct its order resulting in reporting an error in
5.7.13+ release.

FIX
===
The upgrade script should correct the order of the columns
Channel_name and Tls_version in the slave_master_info table
if the order was incorrect.