Bug #25042426: CHANGING WAY OF UNIT TESTS EXECUTION FOR XPLUGIN BY CTEST
The goal of this task is an execution of all xplugin unit test at one
step. At this moment this execution is split into several steps.
In each step, only small groups of tests are being executed.
All unit test are linked in one binary file and it means that this
binary is executed several times with different configuration.
The goal is simplifying of unit tests execution and removing a step of
generation unnecessary configuration files.


Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>
RB: 14489