Bug#24970905: ADD THREAD SANITIZER BLACKLIST
In order to support incrementally fixing issues identified by
Thread Sanitizer (-DWITH_TSAN=1), add support for a blacklist
which can be used to suppress issues yet to be fixed.

The initial version of the blacklist contains suppressions
required to get through main.1st.

Also the patch extends the --sanitize MTR option so that
error logs are scanned for Thread Sanitizer warnings.