Bug#24976885: SDI FILENAMES ARE INCORRECT FOR TABLE NAMES CONTAINING MULTIBYTE CHARACTERS
The sdi filename uses at most 16 characters of the tablename + the object id.
It uses my_mbcharlen(CHARSET_INFO*, uint) to count the number of characters.
The table name is stored in a dd::String_type which means that dereferencing an
iterator yields a (signed) char value.

When the iterator points to the leading byte of a multibyte utf-8 character the
resulting char value is negative. When this value is passed to my_mbcharlen it
is promoted to an unsigned int, which is outside the legal range for unsigned
char (0-255) expected by my_mbcharlen(). my_mbcharlen treats this as an illegal
leading utf8 byte and returns 0.

For sdi file names this meant that the first multibyte character and anything
following it was stripped.

Fix: Make sure inputs to my_mbcharlen are cast to unsigned char.