Bug #24822203: NDB CLUSTER DOESN'T COMPILE WITH GCC6 (DEBUG BUILD)
Problem:
--------

GCC 6 introduces two new types of warnings - misleading indentation
and placement new. These warnings are upgraded to errors in debug
builds.

Fix:
----

The misleading indentation error is fixed while the placement new
one is downgraded to a warning even in debug builds.
Fix of datatype for the offset's in SimpleProperties
list of  instructions for hos to pack/unpack serialized properties
Also one fix in mysqlx_restulset.c which is not yet fixed in 5.7 so
can't be backported