Bug#24926009 'WAITING MAX 119 SEC FOR DISTRIBUTING' ON PARALLEL MYSQLD ACTIONS
'rand()' was used to produce an unique table_id / table_version needed to uniquely
idenify a schema operation being distributed among mysqlds.

This relied on the assumption that rand() would never produced the same numbers
on two different mysqlds.

However, it turns that the random 'seed' is never set when mysqld is init'ed.
According to rand() / srand() documentation this is equivalent of always
setting the seed to '1' - Thus it is *very likely* that the same random numbers
will be produced on all mysqld.

Furthermore, even if seed is set, rand() can produce duplicated. (Not very
likely though)

This patch removes the usage of rand() for producing unique table_id / _version,
and instead use a sequence counter in combination with the node_id of the
coordinator. This guarantees uniquennes until the sequence counter wraps, which
should be sufficient.