Bug#24464763 MOST PERFSCHEMA.*_IO TESTS FAIL IF GIS.SRS RUNS BEFORE THEM ON THE WORKER
This is a test script fix only, to improve test robustness.

Before this fix, some tests failed with spurious diff related to
  Performance_schema_rwlock_instances_lost
in the result file.

This is due to the MTR suite default configuration,
default_mysqld.cnf, which uses a fixed number of instrumented objects:
for some combinations of tests, the number of items actually instrumented
during the test payload can exceed artificial limits set in the test cnf.

The fix is to relax limits on the number of instrumented objects,
and rely on performance schema buffers scaling automatically
based on the load.

Adjusted the test results accordingly.