Bug#24923474 MTR: --SKIP-RPL SHOULD SKIP ALL REPLICATION TESTS AND NONE FROM OTHER SUITES
Problem:
--------
Not all group replication tests are skipped when the --skip-rpl
option is given to MTR. Also, if the absolute path to the MTR
test suite contains the string 'rpl', all tests executed are
identified as replication tests.

Fix:
----
Identify all tests in 'rpl*' and 'group_replication' suites as
rpl_test's and grp_rpl_test's respectively by matching their
suite names. Skip tests which are marked as rpl_test's or
grp_rpl_test's when the --skip-rpl option is given to MTR.

Reviewed-by: Pavan Naik    <pavan.naik@oracle.com>
Reviewed-by: Nuno Carvalho <nuno.carvalho@oracle.com>
RB: 14330