BUG##25055982 : MTR --REPORT-TIMES GIVES TOO HIGH VALUE FOR 'COLLECT' AND 'INIT'
Description :
=============
The --report-times option is meant to give an overview of time
spent on different tasks. Those tasks executed by "workers" like
basic test execution will be summed over all.

But the "Collecting test cases" and "Initialization/cleanup" are
reported with an inflated value if running with --parallel even
though these phases are *not* done by the test workers.

When the workers are started they get their copy of the %time_used
hash where the 'collect' and 'init' already have non-zero values.
When they report this back to the master thread, they will be added
to the total. The end result is that they get multiplied by N + 1
where N is the parallel value.

Fix :
=====
Initialize the '%time_used' hash to value 0 for each worker thread.

Reviewed-by: Bjorn Munch <bjorn.munch@oracle.com>
RB: 14532