BUG#25055190: FLOW-CONTROL APPLIER QUEUE SIZE INCORRECT IN JOINS ON BUSY GROUPS
When a new member joins a group that is busy, one of the statistics
used by the flow-control - applier queue size - may be incorrectly
calculated. This comes from the fact that the queue size is
incremented once the transaction is queued into the relay log, but
it is only decremented when executing if the member is considered
ONLINE. If the member is on RECOVERING status, the queue size is
not decremented.
To fix the above issue, the applier queue counters are only
incremented and decremented when member status is ONLINE.
The certifier queue counter it is always sent, to avoid the
throttling to reach to small values when one or more members are on
RECOVERING, we cap the throttling to 5% of the queues threshold
specified by DBA on
group_replication_flow_control_certifier_threshold and
group_replication_flow_control_applier_threshold options.

A second issue was observed on flow control disable, which was not
enforcing its immediate disable. Now when the low control is disable,
it does reset its used quota which ensures immediate throttling
disable.