Bug#21482542: SOME TESTS FAIL ON WIN WHEN RUN WITH               INNODB_UNDO_TABLESPACES 2
Issue:
------
A couple of tests: main.secure_file_priv_win and
i_main.bug11761752 were failing on Windows with a result
content mismatch because '\'s were being replaced with '/'s.

Fix:
----
The reason for this replacement of slashes was because of
the name of the vardir in the run: innodb-undo-2-debug.
There is a function in MTR, called during --replace_result,
which replaces slashes if it finds a special pattern in the path.
The list of special patterns defined, also contained "undo".
Now, the name of the vardir has been changed.

Reviewed-by: Anitha Gopi <anitha.gopi@oracle.com>
Reviewed-by: Georgi Kodinov <georgi.kodinov@oracle.com>
RB: 14355