Bug#24969065: MEMBER IS NOT MOVING TO ERROR STATE UPON SQL_THREAD FAILURE
After user executed restarts on the group replication applier SQL thread the plugin
could no longer detect when that thread stops due to an error.
This was due to the old detection mechanism that relied on cached thread ids.

To solve this issue, the new detection mechanism simply checks the
channel name to check if it was the group replication applier thread
that failed.