Bug #25042101 SPLIT BINLOG INJECTOR_MUTEX IN TWO, DO REQUIRED CLEANUP
Part 4 of 4

Replace the usage of ndb_schema_share_mutex with the new injector_data_mutex.
ndb_schema_share_mutex was once introduced as the (only) place where we
need a mutex not being as contended as the old injector_mutex.

As we now have the more general injector_data_mutex for such usage,
it only clutters the code having this special ndb_schema_share_mutex.