Bug#25027963: TEMPLATIZE ITEM_CREATE FOR VARIABLE ARGUMENT FUNCTIONS
This patch extends the use of the template introduced in
bugs no 24916689, no 24940128, no 24947913, and no 24947953
where it was used for zero, one, two and three argument
functions, respectively. This patch templatizes the variable
argument functions, and those that posed as such.

This concludes the refactoring, and the last remnants of the
old ways are removed, including the BUILDER, and
GEOM_BUILDER macros.