Bug #24823885: PERFORMANCE REGRESSION WHEN CHANGING CHARACTER SET TO UTF8MB4
Refactoring: Remove char_scanned (and make char_index private); it's not
actually needed, even if we were to add explicit padding for two- or
three-level collations. This means less work done per character,
although the microbenchmarks are pretty much neutral (less than 1% difference),
so this is generally just a cleanup.

Change-Id: I65e1236e4fddccc1ab2c4bc398e2017160b8fc99