Bug#23508203 CANNOT COMPILE OUT FEDERATED STORAGE ENGINE
Building without the feredated storage engine fails.

The root cause is the work around implemented in
  storage/federated/CMakeLists.txt,
which uses file mysys/string.c.

This file was renamed to mysys/string.cc and then to mysys/my_string.cc,
but the makefile was never updated.

This fix adjust the federated makefile to reflect the naming change.