Bug#24624555 DEBUGPRINTDEATHTEST.PRINTEVAL UNIT TEST FAILS ON AARCH64.
Commit 5c2b61d has brought two unit tests (DebugPrintTest and
DebugPrintDeathTest) that use "division by zero" exception to check
the evaluation of args of DBUG_PRINT macro.

About division by zero C99 says in 6.5.5p5: "... if the value of the
second operand is zero, the behavior is undefined".

Also, according to the ARMv8 Architecture Reference Manual if the
divisor is zero, the result of operation is zero and no exception is
generated. Consequently, DebugPrintDeathTest.PrintEval test fails and
DebugPrintTest always passes at least on Aarch64.

Fix: do not use undefined behaviour to test DBUG macros.

The patch is based on a contribution by Alexey Kopytov