Bug #24823885: PERFORMANCE REGRESSION WHEN CHANGING CHARACTER SET TO UTF8MB4
Optimize the hash function; the old one was horrible both in speed and quality.
The new one is significantly faster and slightly less horrible (verified with
SMHasher). Note that we can do this only because the UCA 9.0.0 collations are
not part of a GA release; we cannot go back and do this to other collations
without breaking partitioning.

Microbenchmarks (Skylake 3.4 GHz, optimized, GCC 6.2):

  BM_HashSimpleUTF8MB4   1140 -> 289 ns/iter  [+294.5%]

sysbench goes from 11519 -> 12022 tps (+3.2%).

Change-Id: I6c554110387d927ad1d139c19627c6c51a6aa10c