Bug #24823885: PERFORMANCE REGRESSION WHEN CHANGING CHARACTER SET TO UTF8MB4
Unroll the ASCII fast path to check four bytes at a time. This is a tradeoff;
we lose out on the cases where we have four-byte blocks with mixed ASCII/non-ASCII
(e.g. in text with mostly ASCII but some accents) and on some relatively common
ASCII code points outside the 0x20..0x7e range, such as newlines.

  BM_SimpleUTF8MB4          232 -> 146 ns/iter  [+58.9%]
  BM_MixedUTF8MB4           230 -> 276 ns/iter  [-16.7%]
  BM_MixedUTF8MB4_AS_CS     759 -> 828 ns/iter  [ -8.3%]
  BM_NewlineFilledUTF8MB4   123 -> 231 ns/iter  [-46.8%]
  BM_HashSimpleUTF8MB4      299 -> 306 ns/iter  [ -2.3%]

Change-Id: I64dc2fa06482809cc2e530f2434e5c8890a4edb2