Bug #23596760: FORCE INDEX IS SKIPPED WHILE EXECUTING                SELECT COUNT(*)
Issue:
------
After WL6742, using secondary index for SELECT COUNT(*) has
been replaced by a single call to the handler
(ha_innobase::records()). This function counts the
clustered index and returns the number of records.

This may not be the most optimal in some cases (for example
average row size is large making the clustered index also
large, but with a smaller secondary index). The user can't
use FORCE INDEX syntax to override the call to
handler::records.

Solution:
---------
When FORCE INDEX is specified optimizer doesn't use the
WL#6742 optimization and pick the suggested index.