Bug#25089371: FOLLOWUP AFTER BUG#22874872: MAKE TESTS INDEPENDENT               OF MYSQL.NDB_BINLOG_INDEX
Adjust tests to be agnostic to the existence of the table
'mysql.ndb_binlog_index'.

Create the table at the end of 'scripts/mysql_system_tables.sql'
to make it easier to filter out table ids in the InnoDB sys tables,
since we avoid tables with higher ids that would otherwise be
one off dependeing on the existece of the ndb_binlog_index table.