BUG#24940144 EXPECT.OPEN BEFORE AUTH LEADS TO READ FROM BAD FILEDESCR
Description
===========
X Plugin tries to read from already closed socket, which leads to
"recv" call with socket value set to "-1".

Analyzis and fix
================
The issue was caused by wrong state of "xpl::Client" object which was
not modified in call to "xpl::Client::shutdown_connection" and was causing
additional message-loop iteration in "xpl::Clinet::run" until EBADF
was received.

Reviewed by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>
RB: 14689