Bug #24528148 UNPRIVILEGED USER ABLE TO LOAD COMPONENTS TO MYSQL.COMPONENT
Description: As per wl#4102, there is provision to register components
to mysql.component table but here privileges are not working.
Unprivileged user can able to load/unload these components

Analysis: While loading the components, we are opening the
mysql.component table.
But we are not checking the table access privilege for
the user. This is the causes the bug.

Fix: Added access privilege call before opening the
msql.component table. And moved down the
mysql_dynamic_loader_imp::load and
mysql_dynamic_loader_imp::unload
calls after the open_component_table() because we need
to first check the access privilege before performing
load/unload operations.