Bug#25151440: ALLOW MOVING OR EMPLACING ELEMENTS INTO PREALLOCED_ARRAY
Add variants of insert() and push_back() that accept rvalue references
and move them into place if possible. Also add emplace() and
emplace_back() to construct new elements in place.

This makes it possible to have arrays of types that cannot be copied,
such as std::unique_ptr. This could be useful for example in the
Json_object and Json_array classes in order to simplify memory
management.

Change-Id: Iba54cadbe5baef231d12adf4d4985105dcd82ea1