Bug#25127417 : main.group_by.test fails sporadically on PB2
Problem :
main.group_by.test fails sporadically on PB2 after wl#9489

Solution :
But after detail investigation and discussion among the optimizer team, it was decided that, this particular test has to be retained as Myisam
specific. the following steps are adapted.

A Engine=Myisam has been added in the DDL statement , to retained as Myisam
A column masking at 10 & 11 , has been added for the sporadic explain failures

Reviewed-by: Chaithra <chaithra.gopalareddy@oracle.com>

RB : 14668