Bug#25152435: MOVE CLIENT/SERVER INDEPENDENT FILES FROM SQL-COMMON TO MYSYS
Move my_user.c, my_time.c and pack.c from sql-common/ to mysys/
These files do not need to be compiled separately for client
and server and can therefore be part of the mysys library
and be compiled only once.

Also change them from .c to .cc so that C++ can be used for
future changes in these files.