BUG#25033494 - COMPILATION WARNINGS WITH GROUP REPLICATION PLUGIN - GCS/XCOM
Description
---------------------------

We were warning free when building with Clang (default on OSX/FreeBSD, also
used on Linux) until WL#9053 was pushed.
After the push, I'm counting 312 warnings.

Out of these 150 warnings (there are 2 builds in the log), around 90 belong
to the GCS layer.

How To Repeat:
-------------
Build with Clang

Suggested Fix:
-------------
These issues happen on CLang 3.8/8.0. After a careful analysis, we concluded
that those are false positives on C code.

On XCom sources only, we will inihibit the -Wconditional-uninitialized.