BUG#25134074 - JET/JAG CRASHES WITH MYSQL-GCS
Description:
-----------
From BUG#25098458, during GCS JET tests, one discovered a regression on GCS
caused by a refactoring, which was not discarding messages when the member
was not active within the group.

How To Repeat:
-------------
Please refer to bug#25098458

Suggested Fix:
-------------
Add the check if we are a VOID_NODE_NO when receiving a message.