Bug#20228478 - ON REPLICATION SLAVE ALTER USER FAILING FOR 	       USER WITH SHA256_PASSWORD PLUGIN
Description:
In a replication environment, SET PASSWORD/ALTER USER query
fails to execute on the slave due to slave not being
able to parse the hash string correctly.

Analysis:
When the plain text password is converted to hash and if
the hash string contains a single quote, it results in a
syntax error on the slave thread. This is because the hash
string with quote is written to the binlog without
correctly escaping the special characters. This results in
the slave throwing a syntax error.

Fix:
The fix makes sure that depending on the charset and the
sql-mode of the connection, the hashed password string is
correctly escaped before it gets written to the binlog as
a part of the query. The slave can then correctly parse the
string and can execute the query successfully.