BUG#24643036 BINLOG_SENDER DOES NOT REDUCE SIZE OF SEND BUFFER AS              EXPECTED
Post push fix.

The String object used as network packet by the Binlog sender is not a
real string (it is not always a null terminated string).

The code BUG#24643036 originally pushed did some optimizations in the
packet memory allocation and handling that lead to situations were the
String object buffer didn't filled the last buffer position with a '\0'.

Because of this, some functions handling the buffer as a null terminated
string (packet->c_ptr()) started failing on valgrind tests.

This patch make those functions to pass the pointer to the buffer
without considering it a null terminated string (packet->ptr()).