Bug#24733658: If and friends give wrong type for signed and unsigned integer
The problem here is that the CASE, COALESCE, IF and IFNULL functions merge
signed and unsigned integer data types incorrectly. The value range of
the merged type is not large enough to accomodate all values from the
source expressions. Interestingly, this is only a problem with SELECT,
when the SELECT is used with a CREATE TABLE statement, a correct
data type is generated. This is because the merged field_type is used
to determine the data type in the SELECT case, whereas result_type is used
to determine the data type in the CREATE TABLE case.

The solution is to refactor how the data type of such functions are
generated, and derive the result type from the data type. A new function
Item::aggregate_type() is implemented that loops over all source items
to the function. NULL types are skipped, and if only NULLs are input,
a NULL value is generated. Otherwise, the function Field::field_type_merge
is used to generate the proper data type. However, when the generated
data type is an integer type, and we have both unsigned and signed arguments,
the data type may be adjusted to one with a bigger range. Let's say we have
a signed short int and an unsigned short int as arguments. The value
range for those are -32768..32767 and 0..65535, respectively. The resulting
values are not accomodated in a short int, so we make the resulting type
a signed long int. The range need to be "bumped" only if the largest
unsigned type is the same as the type resulting from field_type_merge().
If the longlong integer type must be bumped, no integer type is wide
enough to hold the resulting values, and a DECIMAL(20,0) is generated.

When implementing Item::aggregate_type(), a few functions were made
redundant and have been deleted. These are item_store_type() (we
no longer aggregate result type, the result type is derived from field
type), agg_result_type() (replaced by field_type_merge()) and
Item_func_if::cache_type_info(). In addition, the resolve_type() functions
of class Item_func_ifnull,  Item_func_if, Item_func_case and
Item_func_coalesce have been made as equal as possible.

This change caused one regression because fields of type BIT and integer
were merged to a string type instead of an integer type. In trunk,
agg_field_type() generated VARCHAR for this combination, but
agg_result_type() generated DECIMAL_RESULT, and this was the chosen
data type. Since a BIT is classified as an integer data type in the
manual, the conversions for BIT and integer types in the array
field_types_merge_rules were changed to numeric types. This caused the
regression to disappear, and no other test failures were introduced.

This will also fix bug#25139420. A test case for this bug is added.